dtype double taken, expected float value error fixed

+++ src/algorithms/deep_q_learning.py
@@ -275,7 +275,7 @@ class DQNAgent:
             q = self.model(torch.cat(samples.state)).gather(1, actions)
 
             # Update the model
-            loss = F.smooth_l1_loss(q, y)
+            loss = F.smooth_l1_loss(q.float(), y.float())
             self._losses.append(loss.item())
             self.optimizer.zero_grad()
             loss.backward()
